LPS-89281 Initialize permissions using existing logic
The DL services expect the serviceContext to be filled in with the
correct permission info. That is handled by the input-permissions tag,
which means that WebDAV requests contain no data at all about
default/allowed permissions.

To make sure we have sane defaults, we need to reset the model
permissions and invoke the deriveDefaultPermissions, which runs the
same logic as the input-permissions tag.