LPS-92284 Remove methods from the API
This whole "View in Context" thing will probably change quite a lot with all the
ongoing work on Display Pages.  Don't expose this on the API until we're sure
we'll need this in the future.

These are the only two componentes that use LayoutFinders, so this duplication
won't propate to other places.