COMMERCE-2420 Calculate "end date" in the past if no end date
We've been only checking whether the calculated end date is after
the start date, and if so, we set it. But the "nullish" end date
value that we were using before sometimes came after the start
date, and thus was used. This change makes sure this never happens
any more.