LPS-109396 Backport Clay v3 packages to frontend-taglib-clay
Commerce is having trouble consuming the Clay v3 packages from
multiple OSGi modules because the shared location for Clay packages
(frontend-taglib-clay) didn't contain any v3 packages on 7.1.x.

Not only is this inconvenient because of the manual configuration
required to install the packages locally, it also means that they wind
up with duplicated versions of the packages, which increases bloat and
in some cases may actually cause things to not work.

About this commit: I added the latest versions of the Clay v3 packages,
same as on current master:

https://github.com/liferay/liferay-portal/blob/55c45e541fc40f21481c3be233892a9e21029e76/modules/apps/frontend-taglib/frontend-taglib-clay/package.json#L3-L38

In theory, there are no semver breaking changes here, but we'll still
want to test this carefully before merging.