LPS-112180 Reapply [role="alert"] on initial refocus
As described here:

https://github.com/antonio-ortega/liferay-portal/pull/113#issuecomment-619120988

> I tested in VoiceOver. When you tab away from the alert, you do indeed
> hear the alert again after one second (but only once). I think the
> right behavior, then, might be to *not* re-add the role on `blur`,
> but rather, on *next* `focus`; ie. only when re-focusing, and with
> the same `relatedTarget` check to make sure that fake event bubbling
> doesn't destroy everything.

So, this commit moves the code from `blur` to `focus`. Re-testing shows
that it gives the desired result, at least with VoiceOver. Reads out the
alert text when the alert first appears; doesn't re-read it just because
you tab around inside the alert; but does re-read it exactly once when
you tab back into the alert from the outside.