LPS-117924 Simplify yarn workspace globs
This is the re-do of:

https://github.com/brianchandotcom/liferay-portal/pull/91892 (master PR)

for 7.1.x.

As explained in that PR, a common question we get in Slack asks why the
"liferay-npm-scripts" executable can't be found at build time, and the
problem is usually a missing glob in the Yarn workspaces list. So the
idea here is to eliminate that class of problem by making our globs much
looser by default. We were initially more conservative out of caution,
but performance seems fine with the broader globs and the build is
unchanged.

I took the "compare.js" script from that pull:

https://gist.github.com/wincent/fea9f37abe2261526cf5b9a8191296ee

And modified it for 7.1.x:

https://gist.github.com/wincent/3a9f7a91ea8380f21b3ef1ae72c7ae6a

It printed:

    Common: 64, Missing: 0, Extra: 3

    Extra:

    - apps/frontend-theme/frontend-theme-unstyled/package.json
    - dxp/apps/osb/osb-faro/osb-faro-theme/package.json
    - dxp/apps/osb/osb-faro/osb-faro-web/package.json

Now, the osb-faro stuff is in a world of its own (without yarn and still
using package-lock.json files, and having custom devDependencies etc),
and I don't think we want to add workspaces for that (which would drag
all of the dependencies into the core).

But we can add frontend-theme-unstyled to the workspaces list (it
participates in the workspaces mechanism on 7.2.x and master).

So the after tweaking the globs list we get:

    Common: 64, Missing: 0, Extra: 1

    Extra:

    - apps/frontend-theme/frontend-theme-unstyled/package.json

ie. identical workspace definition before and after, except for that one
addition.

The lockfile on this branch is stale, so a `yarn install` ends up
including then lockfile changes you see in this commit.