LPS-117924 Deduplicate yarn.lock
We are normally pretty scared of touching the yarn.lock on the older
branches, but seeing as I already touched it necessarily in the last
commit I think we may as well take the opportunity to deduplicate it; in
the end, the changes are pretty minimal and I think the risk is low.

Done using `npx yarn-deduplicat yarn.lock` from the "modules/"
directory.