LPS-97451 Comparing plid's avoids necessary update of preferences for layout
When a fragment configuration is changed in edit view, the portlet preferences
point to the control panel plid of the group (7.1.x only). To update the portlet
preferences of the layout the two plids will necessarily be different.

Original problem LPS-95512 is still taken care of by comparing groupId's because
live and staging portlet preferences with the same portletId will differ there.